[arange] set dtype='int64' explicitly. (#6071)

+++ paddlenlp/transformers/clipseg/modeling.py
@@ -237,7 +237,9 @@ class CLIPSegTextEmbeddings(nn.Layer):
         self.position_embedding = nn.Embedding(config.max_position_embeddings, embed_dim)
 
         # position_ids (1, len position emb) is contiguous in memory and exported when serialized
-        self.register_buffer("position_ids", paddle.arange(config.max_position_embeddings).expand((1, -1)))
+        self.register_buffer(
+            "position_ids", paddle.arange(config.max_position_embeddings, dtype="int64").expand((1, -1))
+        )
 
     def forward(
         self,
